Good_Bad_Customers_for_Granting_Credit_IIT-R
Problem: The banks with the intent of credit card were more focused on the number of customers using their credit service but the drawback of them not being able to pay back the credit in time was an issue that soon followed, a system was in need to effectively decide the credit limit to be allowed to a person based on his previous credit history.

Objective: Build a classification model using logistic regression to predict the credibility of the customer, in order to minimize the risk and maximize the profit of a bank. Compare it with one another algorithm of your choice. Dataset:Source(https://www.kaggle.com/uciml/default-of-credit-card-clients-dataset)