Project: Good Bad Customer for Granting Credit
Project Brief: The banks with the intention of increasing the number of customers using their credit service were more focused on growing this segment, but the drawback was that many customers who were given credit were not able to pay back the credit on time, and this became a concerning issue soon. A system was required to effectively decide which customers were worthy of being granted credit based on his/her previous credit history.

Objective: Build a classification model using Logistic Regression to predict if a customer should be granted credit or not, so that the bank can minimize its risk and maximize its profit. Compare it with another algorithm of your choice. Dataset:Source(https://www.kaggle.com/uciml/default-of-credit-card-clients-dataset)